♻️ commands.rs と lib.rs の不揃いを直す
